improve the status bar information for bodies in SceneViews

+++ src/Util/SceneShape.h
@@ -564,19 +564,19 @@ public:
         lineVertices_.push_back(v1);
     }
 
-    void setLineWidth(double width) { lineWidth_ = width; }
+    void setLineWidth(float width) { lineWidth_ = width; }
 
     /**
        The default value of this is zero and the current system value is used then.
     */
-    double lineWidth() const { return lineWidth_; }
+    float lineWidth() const { return lineWidth_; }
 
 protected:
     SgLineSet(const SgLineSet& org, SgCloneMap& cloneMap);
 
 private:
     SgIndexArray lineVertices_;
-    double lineWidth_;
+    float lineWidth_;
 };
 typedef ref_ptr<SgLineSet> SgLineSetPtr;
 
@@ -589,12 +589,12 @@ public:
     virtual void accept(SceneVisitor& visitor);
     const Vector4f& color() const { return color_; }
     void setColor(const Vector4f& color) { color_ = color; }
-    void setLineWidth(double width) { lineWidth_ = width; }
-    double lineWidth() const { return lineWidth_; }
+    void setLineWidth(float width) { lineWidth_ = width; }
+    float lineWidth() const { return lineWidth_; }
 
 private:
     Vector4f color_;
-    double lineWidth_;
+    float lineWidth_;
 };
 typedef ref_ptr<SgOutlineGroup> SgOutlineGroupPtr;
 }
